+++ src/psTerminalPerfCounter/psTerminalPerfCounter/Private/New-CounterConfigurationFromJson.ps1
@@ -11,9 +11,9 @@
         [Parameter()]
         [bool]              $isRemote,
         [Parameter()]
-        [string]            $computername,
+        [string]            $ComputerName,
         [Parameter()]
-        [pscredential]      $credential
+        [pscredential]      $Credential
     )
 
     $PerformanceCounters = [System.Collections.Generic.List[psTPCCLASSES.CounterConfiguration]]::new()
@@ -35,14 +35,14 @@
             $CounterConfig.colorMap,
             $CounterConfig.graphConfiguration,
             $isRemote,
-            $computername,
-            $credential,
+            $ComputerName,
+            $Credential,
             $counterMap
         )
 
         $PerformanceCounters.Add($CounterConfiguration)
     }
 
-    return $PerformanceCounters
+    return , $PerformanceCounters
 
 }