+++ src/psTerminalPerfCounter/psTerminalPerfCounter/Public/Get-tpcPerformanceCounterInfo.ps1
@@ -36,12 +36,12 @@ function Get-tpcPerformanceCounterInfo {
         [string] $SearchTerm
     )
 
-    $Result = @()
+    $Result = [System.Collections.Generic.List[object]]::new()
     $param  = @{}
 
     if ( $PSCmdlet.ParameterSetName -eq 'Remote' ) {
 
-        $param.Computername = $computername
+        $param.Computername = $ComputerName
 
         if ( $null -ne $Credential ) {
             $param.Credential = $Credential
@@ -51,7 +51,7 @@ function Get-tpcPerformanceCounterInfo {
     try {
 
         if ( $PSCmdlet.ParameterSetName -eq 'Remote' -and -not $(Test-Connection -ComputerName $ComputerName -Count 1 -Quiet) ) {
-            THROW "Remote computer $computername not reachable. Aborting"
+            THROW "Remote computer $ComputerName not reachable. Aborting"
         }
 
         # is ID
@@ -88,13 +88,13 @@ function Get-tpcPerformanceCounterInfo {
                         } | Select-Object -Unique
                     }
 
-                    $Result = [PSCustomObject]@{
+                    $Result.Add([PSCustomObject]@{
                         ID          = $SearchTerm
                         CounterSet  = $SetName
                         Path        = $PathName
                         SetType     = $SetType
                         Instances   = $($counterInstancesValues -join ', ')
-                    }
+                    })
 
                 } else {
                     Write-Host "Could not resolve ID '$SearchTerm' to counter names." -ForegroundColor Red
@@ -157,13 +157,13 @@ function Get-tpcPerformanceCounterInfo {
 
                         }
 
-                        $Result += [PSCustomObject]@{
+                        $Result.Add([PSCustomObject]@{
                             ID          = $CompositeId
                             CounterSet  = $CounterSet.CounterSetName
                             Path        = $CounterPath
                             SetType     = $SetType
                             Instances   = $($counterInstancesValues -join ', ')
-                        }
+                        })
 
                     } catch {
                         # Warning instead of Error to avoid stopping the entire loop for one bad counter